pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server for Morpho Protocol"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.9"
websockets = "^12.0"
requests = "^2.31.0"
typing-extensions = "^4.9.0"
web3 = "^6.10.0"
python-dotenv = "^1.0.0"
sseclient-py = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
